Moved OpcodeBase to shared folder
